Add footer section
The final section of the webpage has been finished which concludes this activity for the
foundations course of The Odin Project. This webpage is far from finished however, As I
intend to improve the design and make it more personalized in the future. JavaScript can
be added once I get comfortable with it.